% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ordination.R
\name{extract_ordination_axes}
\alias{extract_ordination_axes}
\title{Extract site scores and axis labels from an ordination object}
\usage{
extract_ordination_axes(ord, n_axes, verbose = TRUE)
}
\arguments{
\item{ord}{An ordination object.}

\item{n_axes}{The number of axes wanted.}

\item{verbose}{Logical. If TRUE, report the eigenvalue convention used.}
}
\value{
A list with \code{scores} (a matrix of site scores, samples in rows),
\code{labels} (axis labels including percent variation where available) and
\code{n_axes} (the number of axes actually available).
}
\description{
Extract site scores and axis labels from an ordination object
}
\keyword{internal}
