% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ordination.R
\name{ordi_methods}
\alias{ordi_methods}
\title{Ordination methods accepted by the star plot functions}
\usage{
ordi_methods()
}
\description{
These are the methods \code{phyloseq::ordinate()} understands. Note that "PCA" is
not among them: an unconstrained RDA is the phyloseq spelling of a PCA.
}
\keyword{internal}
